prep for transit combo
